add 2020 VolunteerController with routes and fix imagesView path error
